Perf upgrade (#100)
* move away from singular bpf read/deletes to bpf batch read/delete
* migrate from hashmap to ringbuf for sk_props
* memory optimizations / struct compactions
* expose max_sock_props
* chore: Update eksbuild tag to v1.1.4-eksbuild.1 and chart version to 1.1.4
* add SYS_ADMIN when open_metrics is on
---------

+++ nfm-controller/src/reports/publisher_amp.rs
@@ -418,7 +418,7 @@ mod tests {
         let mut report = NfmReport::new();
         let context = SockContext {
             is_client: false,
-            address_family: libc::AF_INET as u32,
+            address_family: libc::AF_INET as u16,
             local_ipv4: 16909060,
             remote_ipv4: 84281096,
             local_ipv6: [0; 16],
@@ -520,7 +520,7 @@ mod tests {
 
         let context = SockContext {
             is_client: false,
-            address_family: libc::AF_INET as u32,
+            address_family: libc::AF_INET as u16,
             local_ipv4: 16909060,
             remote_ipv4: 84281096,
             local_ipv6: [0; 16],
@@ -608,7 +608,7 @@ mod tests {
 
         let context = SockContext {
             is_client: false,
-            address_family: libc::AF_INET as u32,
+            address_family: libc::AF_INET as u16,
             local_ipv4: 16909060,
             remote_ipv4: 84281096,
             local_ipv6: [0; 16],
@@ -662,7 +662,7 @@ mod tests {
 
         let context = SockContext {
             is_client: false,
-            address_family: libc::AF_INET as u32,
+            address_family: libc::AF_INET as u16,
             local_ipv4: 16909060,
             remote_ipv4: 84281096,
             local_ipv6: [0; 16],
@@ -799,7 +799,7 @@ mod tests {
 
         let context = SockContext {
             is_client: false,
-            address_family: libc::AF_INET as u32,
+            address_family: libc::AF_INET as u16,
             local_ipv4: 16909060,
             remote_ipv4: 84281096,
             local_ipv6: [0; 16],
@@ -892,7 +892,7 @@ mod tests {
 
         let context = SockContext {
             is_client: false,
-            address_family: libc::AF_INET as u32,
+            address_family: libc::AF_INET as u16,
             local_ipv4: 16909060,
             remote_ipv4: 84281096,
             local_ipv6: [0; 16],
